update cors and refactor
