+++ src/main/java/com/huylam/realestateserver/config/WebConfig.java
@@ -1,6 +1,7 @@
 package com.huylam.realestateserver.config;
 
 import org.springframework.context.annotation.Configuration;
+import org.springframework.web.servlet.config.annotation.CorsRegistry;
 import org.springframework.web.servlet.config.annotation.PathMatchConfigurer;
 import org.springframework.web.servlet.config.annotation.WebMvcConfigurer;
 
@@ -10,5 +11,15 @@ public class WebConfig implements WebMvcConfigurer {
   @Override
   public void configurePathMatch(PathMatchConfigurer configurer) {
     configurer.setUseTrailingSlashMatch(true);
+  }
+
+  @Override
+  public void addCorsMappings(CorsRegistry registry) {
+    registry
+      .addMapping("/api/v1/auth/**")
+      .allowedOrigins("http://127.0.0.1:5173")
+      .allowedMethods("GET", "POST")
+      .allowedHeaders("*")
+      .allowCredentials(true);
   }
 }
